Added NO_SYMBOLS_CFLAGS, etc. to allow stripping symbols when --disable-symbols is specified and to ensure that production mode (where symbol stripping is usually a part of the FLAGS) doesn't conflict with --enable-symbols. This will allow better (though still limited) debugging of production/optimized code.
The PGI, XLC, Intel, and Solaris files were not modified and
do not generate the symbol removal flag at this time. They'll
be updated in future check-ins. In the meantime, production
mode on those platforms still usually strips symbols.

This check-in also updates the Cygwin-specific files to use
the new autotools FLAGS and option scheme when uncommon
(non-gnu, etc.) Fortran compilers are used.

Tested on: 64-bit Ubuntu 15.10 (Linux 4.2.0 x86_64) gcc 5.2.1
           autotools serial (prod/debug w/ and w/o symbols)
           symbol presence/absence checked with objdump


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@29129 dab4d1a6-ed17-0410-a064-d1ae371a2980